refactor(registration-form): replace the old form with a new Formik one
