+++ client/src/App.js
@@ -7,12 +7,11 @@ import {
   Dashboard,
   Error404,
   Error500,
-  Form,
   Loading,
   PrivateRoute,
 } from '@app/components';
 import '@app/App.css';
-import RegistrationPage from './components/Form/Form-new';
+import RegistrationPage from './components/Form/Form';
 
 function App() {
   const { isLoading } = useAuth0();
@@ -25,7 +24,7 @@ function App() {
     <div className="App">
       <Router>
         <Switch>
-          <Route path="/" exact component={Form} />
+          <Route path="/" exact component={RegistrationPage} />
           <Route path="/registration" exact component={RegistrationPage} />
           <Route path="/confirmation" exact component={Confirmation} />
           <PrivateRoute path="/dashboard" component={Dashboard} />
